Factor network view out of the main app
